requirements.txt
Django==2.2.7
djangorestframework==3.10.3
dj-database-url==0.5.0
gunicorn==20.0.2
requests==2.22.0

flake8==3.7.9
pep8==1.7.1
